fix(export): un-double a backslash before merged.rs re-escapes it in FILE_SCHEMA
Verified the read-side gap raised in weekend review discussion around #2405/
#2408 (write-side backslash doubling for the STEP and merged exporters): does
merged.rs's read side un-double `\\` before interpreting a STEP string
literal, and does write->read->write preserve it exactly?

Ground truth, mapped site-by-site in merged.rs:
- rewrite_refs (DATA-section entity lines): byte-opaque copy. It tracks
  in/out-of-string state via the same doubled-apostrophe toggle trick as
  the core scanner, purely to protect `#` references from being rewritten
  inside a literal; it never decodes or re-encodes string content. Proven
  clean with a round-trip test carrying a doubled apostrophe, a doubled
  reverse solidus, and a `\X2\...\X0\` UCS-2 directive in one DATA-section
  string -- passes through byte-for-byte. This closes the specific worry
  from the review (a decode-without-reencode, or interpret-without-decode,
  path corrupting DATA content); it does not exist.
- detect_schema (HEADER FILE_SCHEMA field): extracts a RAW, still-escaped
  slice and feeds it straight into escape() on write, which doubles `\`
  again. This is a genuine half-decode/re-encode mismatch, but only for a
  schema label containing '\' or "'" -- no real file's schema (IFC2X3,
  IFC4, IFC4X3_ADD2, ...) ever does. It was already flagged as a known,
  deliberately-unfixed bound in #2408's commit message.

Fixed the backslash half here: detect_schema now un-doubles `\\` in the
extracted text before returning, so escape() re-doubles the correct,
already-decoded value instead of doubling on top of doubling. No-op for
every real schema label. The doubled-apostrophe half of the same
raw-extraction gap is a separate, already-tracked quote-scanning defect
(naive first-two-quotes search, not a doubling mismatch), addressed by
unifying onto step_text::detect_schema() on the unmerged
fix/merged-detect-schema-quote2 branch -- out of scope here to avoid
duplicating that in-flight work.

merged.rs's test module moved to merged_tests.rs (the `_tests.rs`-suffix
house convention, same as step_text.rs) so the two new round-trip tests
don't push the production module over the 400-line ratchet.

Verification: RED (assert_eq! against the pre-fix un-doubled value) then
GREEN on the schema fix; the DATA-section round-trip test is a pin (already
green on main) verified to catch a mutation that reintroduces an in-string
decode. cargo test -p ifc-lite-export --lib: 184 passed / 37 failed, before
and after (the 37 are pre-existing tests/models/ fixture-path failures in
this worktree, identical count both times). cargo test -p ifc-lite-processing
--test module_size_ratchet: 4/4 passed. cargo clippy --workspace --exclude
ifc-lite-wasm --all-targets -D warnings: clean.

Rust-only change (precedent a4437d098): no changeset.
